Updated: Valve check state, AC-> VI communication
Addition of automated valve cycle to the valve check state. Using command cyclevalves inside Valve check state will allow to cycle from sov1 to sov9  checking that the correct gpio is being turn on and communicating the GPIO status as we cycle the valves to the Valve indicator. Added communication between AC and VI using USART 3 .